+++ src/eye_boof/visualization.clj
@@ -2,6 +2,7 @@
   "Provides functions to visualize images using JFrame."
   (require [eye-boof.core.io :refer [resource->buff-image]])
   (:import boofcv.gui.image.ShowImages
+           boofcv.gui.ListDisplayPanel
            boofcv.gui.binary.VisualizeBinaryData
            boofcv.struct.image.ImageBase))
 
@@ -9,7 +10,9 @@
 (set! *unchecked-math* true)
 
 (defprotocol ViewImageWindow
-  (show-image [img] "Shows an image on a JFrame window."))
+  (show-image [img]
+    "Shows an image on a JFrame window. If a collection of images is passed, a
+list panel is generated."))
 
 (extend-protocol ViewImageWindow
   java.awt.image.BufferedImage
@@ -18,7 +21,23 @@
   ;;
   ImageBase
   (show-image [img]
-    (show-image (resource->buff-image img))))
+    (show-image (resource->buff-image img)))
+  ;;
+  clojure.lang.IPersistentCollection
+  (show-image [imgs]
+    (ShowImages/showWindow
+     ^ListDisplayPanel (reduce (fn [^ListDisplayPanel ldp ^ImageBase i]
+                                 (.addImage ^ListDisplayPanel ldp i "Image")
+                                 ldp)
+                               (ListDisplayPanel.)
+                               imgs)
+     "Images")))
+
+(defn show-grid
+  "Shows images in a grid pattern."
+  [ncols & imgs]
+  (ShowImages/showGrid
+   ncols "Images" (into-array (map resource->buff-image imgs))))
 
 (defn render-binary
   "GrayU8 -> boolean -> BufferedImage